factor out selectUtils
